[move-lang] Typed macros: parsing and type checking
(This is a port of [unsubmitted PR #1005](https://github.com/diem/diem/pull/10005) from the Diem repo to the Move repo. Mentioned PR was submitted before Move was split off Diem. For the historical discussion see the original version. The port has been create by (a) calling `git format-patch` in the Diem repo (b) calling `git am -3 < <patch-file>` in the Move repo, so it should be save following the same schema as git merge does.)

This PR realizes the 1st step towards an alternative, high fidelity implementation of user level macros, in contrast to compiler macros suggested in [Diem PR #9930](https://github.com/diem/diem/pull/10005).

In this model, macros can be declared using a new style of function declaration:

```
    public macro foreach<T>(v: &vector<T>, action: |&T|) {
        let i = 0;
        while (i < Vector::length(v)) {
            action(Vector::borrow(v, i));
            i = i + 1;
        }
    }
```

This PR implements parsing and type checking of this style of macros up to the typing AST. Macro expansion is not yet implemented and intended to be added in a followup.

The biggest change is enabling function types, lambda expressions, and calls of function values for naming and typing. Those had been supported before for the spec language, but weren't supported in move-lang beyond expansion. In the design, we represent function types by a new BuiltinTypeName_::Fun which allows to minimize changes to the code, as no new type variant needs to be introduced, and existing functionality for inference et. al can be adopted with minimal changes.

A new test case has been added at [lambda.move](https://github.com/wrwg/move/blob/macros_alt/language/move-lang/tests/move_check/typing/lambda.move) which shows usage and intends to cover all type check failure scenarios.

After this PR, to complete this style of macros, we still need to:

- Implement actual expansion of macros. This is intended to happen in a new phase after typing.
- Implement representation of macros in compiled, persisted modules. We need to be able to fetch the macro definition from an imported, precompiled module.
- Implement techniques (in the prover) to refer to specifications of passed lambda arguments to macros. (See PR #9978 for what we want to do here.)